% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bucaramanga.R
\docType{data}
\name{Bucaramanga}
\alias{Bucaramanga}
\title{A dataframe of Bucaramanga station from IDEAM (Colombia)}
\format{
A dataframe with 21 information columns of of Bucaramanga station, the
variable downloaded is precipitation from 1999-01-01 to 2018-12-31.
}
\usage{
Bucaramanga
}
\description{
A dataframe of Bucaramanga station from IDEAM (Colombia), data downloaded of precipitation
from DHIME (IDEAM, Colombia).
}
\keyword{datasets}
