% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideam2annual.R
\name{ideam2annual}
\alias{ideam2annual}
\title{function for calculate monthly values of prepared data from DHIME (IDEAM, Colombia)}
\usage{
ideam2annual(x, param = NULL, na.rm = NULL)
}
\arguments{
\item{x}{dataframe of prepared data of a station from DHIME (IDEAM, Colombia)}

\item{param}{default value is sum, sum is for precipitation and evaporation, mean for temperature and relative humidity.}

\item{na.rm}{default value is TRUE, TRUE for consider NA, FALSE for not consider NA.}
}
\description{
function for calculate monthly values of prepared data from DHIME (IDEAM, Colombia),
apply to precipitation, minimum and maximum temperature.
}
\author{
Geomar Perales Apaico
}
