% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideam2monthly.R
\name{ideam2monthly}
\alias{ideam2monthly}
\title{function for calculate monthly values of prepared data from DHIME (IDEAM, Colombia)}
\usage{
ideam2monthly(x, param = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{dataframe of prepared data of a station from DHIME (IDEAM, Colombia).}

\item{param}{default value is sum, sum is for precipitation and evaporation, mean for temperature and relative humidity, max for maximum values.}

\item{na.rm}{default value is TRUE, TRUE for consider NA, FALSE for not consider NA.}
}
\description{
function for calculate monthly values of prepared data from DHIME (IDEAM, Colombia),
apply to precipitation, minimum and maximum temperature.
}
\examples{
data <- ideamprep(Bucaramanga)
ideam2monthly(data)

}
\author{
Geomar Perales Apaico
}
