% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideaminfo.R
\name{ideaminfo}
\alias{ideaminfo}
\title{function for obtain information about data downloaded of station from DHIME (IDEAM, Colombia)}
\usage{
ideaminfo(x)
}
\arguments{
\item{x}{dataframe of a station from DHIME (IDEAM, Colombia).}
}
\description{
function for obtain information about data downloaded of station from DHIME (IDEAM, Colombia),
apply to precipitation, minimum and maximum temperature.
}
\examples{
ideaminfo(Bucaramanga)

}
\author{
Geomar Perales Apaico
}
