% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideamloc.R
\name{ideamloc}
\alias{ideamloc}
\title{function for obtain location of data downloaded of station from DHIME (IDEAM, Colombia)}
\usage{
ideamloc(x)
}
\arguments{
\item{x}{dataframe of a station from DHIME (IDEAM, Colombia).}
}
\description{
function for obtain location of data downloaded of station from DHIME (IDEAM, Colombia),
apply to precipitation, minimum and maximum temperature.
}
\examples{
ideamloc(Bucaramanga)

}
\author{
Geomar Perales Apaico
}
