% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideamna.R
\name{ideamna}
\alias{ideamna}
\title{function for calculate days without information by month of prepared data of station from DHIME (IDEAM, Colombia)}
\usage{
ideamna(x, type = NULL)
}
\arguments{
\item{x}{dataframe of prepared data of a station from DHIME (IDEAM, Colombia).}

\item{type}{default value is NULL, NULL is for days without information by month, inv is for days with information by month.}
}
\description{
function for calculate days without information by month of prepared data of station from DHIME (IDEAM, Colombia),
apply to precipitation, minimum and maximum temperature.
}
