% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideamprep.R
\name{ideamprep}
\alias{ideamprep}
\title{function for prepare data from DHIME (IDEAM, Colombia)}
\usage{
ideamprep(x, level = NULL)
}
\arguments{
\item{x}{dataframe of a station from DHIME (IDEAM, Colombia).}

\item{level}{information level of dataframe, default information is standard.
standard: dataframe of date, station, parameter and parameter values
basic: dataframe of date and parameter values
advanced: dataframe of date, station, parameter, latitude, longitude, height and parameter values}
}
\description{
function for prepare data from DHIME (IDEAM, Colombia), apply to precipitation,
minimum and maximum temperature.
}
