% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideamsummary.R
\name{ideamsummary}
\alias{ideamsummary}
\title{function for obtain information summary of data downloaded of station from DHIME (IDEAM, Colombia)}
\usage{
ideamsummary(x)
}
\arguments{
\item{x}{dataframe of a station from DHIME (IDEAM, Colombia).}
}
\description{
function for obtain information summary of data downloaded of station from DHIME (IDEAM, Colombia),
apply to precipitation, minimum and maximum temperature.
}
\examples{
ideamsummary(Bucaramanga)

}
\author{
Geomar Perales Apaico
}
