% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideamprep.R
\name{ideamprep}
\alias{ideamprep}
\title{function for prepare data from DHIME (IDEAM, Colombia)}
\usage{
ideamprep(x)
}
\arguments{
\item{x}{dataframe of a station from DHIME (IDEAM, Colombia).}
}
\description{
function for prepare data from DHIME (IDEAM, Colombia), apply to precipitation,
minimum and maximum temperature.
}
\author{
Geomar Perales Apaico
}
